Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/custom-elements/range-and-constructors-expected.txt


FAIL Range.cloneContents should invoke constructor in tree order assert_array_equals: property 0, expected "root-0" but got "root-0-0"
FAIL Range.extractContents should invoke constructor in tree order assert_array_equals: property 0, expected "root-0" but got "root-0-0"
